requirements.txt
Django>=6.0,<7
channels>=4.0
channels-redis>=4.2
daphne>=4.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg[binary]>=3.2
python-dotenv>=1.0
drf-spectacular>=0.27
